[IMP] auth_signup: hide unusable field from users
The token field is a technical data that the other users are not able to use.
It may be confusing for users to see token on the user interface.
Still show it to administrator for debug reasons.

(cherry picked from commit 7501691da9a9dae832b7d6e78c6a5f57fc983b5c)

# Conflicts:
#	addons/auth_signup/models/res_users.py