[IMP] account: not allow to have different company in move line and account move
Purpose of the commit is when multi-company is active and company A is a
parent of company B.
When a user creates an account move with a Journal of company A and adds
account move lines they are created with company A. But now user changes
the journal and select journal of different company and save the record system
allows having the different company in account move and account move line.

After this commit when account move and account move line has different
company and tries to save, it will show a warning.

task-1970459

closes odoo/odoo#33864

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>