[FIX] web_editor: remove unused template
The wysiwyg.widgets.image.existing template was not used and
reintroduced in a refactoring. It is now gone as it should be.

Commit where the template is removed with its only usage:
(in editor.xml as web_editor.dialog.image.existing)
78a30d44cf8f1ce0d4d3fc53481a87eebeaea088

Commit where the template is reappearing refactored but not used:
f296992317e96562c66bd7ad59a5080d6c551ed5

Part of #33983