[FIX] base: Remove label from views if user groups forbid the field
Without this patch, if a view defines a `<label for="forbidden_field"/>` anywhere, the web client tries to fetch the field data, but since the user has no access, it gets an error.

Now, if the label targets a forbidden field, it gets removed from the view, as happens with the related field. The client doesn't know of that field's existence and no errors happen.

closes odoo/odoo#38868

X-original-commit: a41aa063fd24775aa23b5264129b0e5344dc367f
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>