[FIX] mail: use consistent placeholder
When generation the message. add the name of the channel with a
placeholder. In some languages like Japanese, the order of words is
different and adding the name at the end makes an incorrect sentence,
it was not possible to properly translate it.

(cherry picked from commit 4827ef3ab24569db09513378e664d289b427fae3)