[IMP] base: make datetime coherent with dateutil
As we do not need all the datetime module, we want to restrict the
functions (as we do for dateutil). Also centralise handling of this
and apply it by default, no reason for there to be divergences.

Task 2179629

(cherry picked from commit cd32b0c5e6e9fb1ee67a17266eb388be99786f09)