[FIX] account,hr_expense: Fix unbalanced journal entry created by expenses
Suppose the journal entry contains 2 identical expenses using the same tax in a locked period. Your journal entry looks like:
- exp1		amount
- tax_exp1:	tax_amount
- total		- amount - tax_amount
- exp2		amount
- tax_exp2	tax_amount
- total		- amount - tax_amount
This is because the expense are creating manually their tax lines.
When posting, the "Cannot create unbalanced journal entry" is raised because the date is automatically set to the next available date after the lock date and then, _onchange_currency is triggered on the journal entry.
Because tax_base_amount is expressed in the company's currency and the rate has changed, _recompute_tax_lines is called with recompute_tax_base_amount=True.
The tricky part is the journal entry always group tax lines together as much as possible but don't update the tax amount due to the recompute_tax_base_amount parameter.
Before this commit, one of the tax lines was removed and the other wasn't updated leading to the unbalanced journal entry. After this commit, there is no modification except the tax_base_amount that is made with the recompute_tax_base_amount parameter.

closes odoo/odoo#64318

Co-author: elmeriniemela (https://github.com/odoo/odoo/pull/63985)
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>