[bt#19975] Update pdfjs from origin/14.0
git checkout d3bb55afb048e4690f3b8d91 ./addons/web/static/lib/pdfjs

I suppose pdfjs isn't update very often on the project branch. In this case, it
showed a "bad XRef entry" instead of the PDF. Evince on the other hand was
perfectly capable of rendering it.

Not sure if this is the way to go. I suppose it must be tested carefully and
the component is very isolated. Let's see.