[FIX] account_edi: fix _compute_edi_format_ids on journal
When computing the edi_formats on a journal, keep the ones
that are already checked (if they are compatible with this journal).

The bug occurs when migrating a DB: if a new edi_format is created,
the `create` will call the `_compute_edi_format_ids` on all journals.
Thus, all edi_formats will be reset since nothing keeps track of the
already checked edi_formats. This PR fixes this.

closes odoo/odoo#102281

Signed-off-by: Laurent Smet <las@odoo.com>