[FIX] account,l10n_ch: taxes update fix when no template changed
When no new template are created in the update of taxes migration method
the execution lead to an an error in the `_process_taxes_translations` method.
Fixed a similar issue in l10n_ch migration that enables the taxes afterward
and could possibly cash in the same way if no taxes are created.

opw-3305302

closes odoo/odoo#121184

Signed-off-by: John Laterre (jol) <jol@odoo.com>