[FIX] Mail `chatter`: fix followers dropdown
Before this commit, the dropdown menu lacked spacing at the top and
bottom when a user clicked on "Followers" in the chatter.

This commit addresses the issue by adding the necessary spacing and also
adjusting the hover state background color (when hovering followers
details and actions) to improve the contrast with the text color.

task-3338438
part of task task-3326263

closes odoo/odoo#129649

X-original-commit: ead5141a90122e3556da6d124c01f4198fa4bfb5
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>