[FIX] account: avoid infinite loop when creating taxes of type 'group'
When updating taxes, creation of taxes of type 'group' does not
work correctly if either a child or the parent itself has been deleted.
Indeed the creation of _generate_tax() method suppose that both the parent
and the child needs to be created in the same batch while we can have
databases where one of those already exists and will go in the 'update'
and not 'creation' part of the script.

closes odoo/odoo#130911

X-original-commit: d20f2e2b7f1b03d786e1a2298aae4d49b7923eae
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>