[IMP] account_edi_ubl_cii: improve import of XML invoice with errors
Sometimes the user needs to import XML invoices (e.g. factur-x)
containing lines with the following problem:
The billed quantity of the line is set to 0 but the total of the line
(plus/minus allowances) is not 0.

The imported version (in odoo) of this line will have quantity 0 and thus a
total of 0 (since the total will be computed in odoo and not stored
directly in the database).

After this commit the quantity of these line will be calculated
such that the total computation yields the correct total.

To do this an already existing similar "exception" in the parsing
logic was extended to deal with this problem.

closes odoo/odoo#137833

Signed-off-by: Laurent Smet (las) <las@odoo.com>