[FIX] pos_sale: not pos_groupable uom with float quantity pos sale
Issue:
======
When we settle an order that has a product with uom not pos_groupable
and quantity isn't an integer for example 3.5 it will add 4 lines each
with quantity = 1 which is not correct.

Steps to reproduce the issue:
=============================
- Create a product with a not pos_groupable uom (kg for example)
- Create a sale order with that prouct and add a not integer quantity
- Go to pos and settle the order
- You will have 4 lines added each with quantity = 1

Origin of the issue:
====================
The quantity was hardcoded = 1

Solution:
=========
I addeed order lines equal to the integer part of quantity and one last
order line that have the rest quantity.

opw-3579225

closes odoo/odoo#143566

X-original-commit: 3db0515959861e65f05b572d4827692ed00a1925
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>