[FIX] test_main_flows: edit record after Save & New
Since the commit afdcf9a5d5cd0d25b2ba61167d3ae6726a97d554, the mobile main flow tour
fail sometimes. This is due to the fact that the tour try to edit the record
before the new record is displayed.

The solution is to wait for the new record to be displayed before trying to edit it.

closes odoo/odoo#142231

X-original-commit: f10e7e01e41e6a5d589ae9e1a5e92d9ac9994686
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Francois Georis (fge) <fge@odoo.com>