[FIX] mail: improve activity state search method
The method _search_activity_state is trigger each time someone click
on the progress bar of lead kanban view and at many other occasion.

This search can take up to 2 seconds + a seconds query id in a very long
list of ids that take also 2 seconds to run.

This PR improve the perf of this search by removing join with huge table
(res_users and res_partner) by storing the tz on the activity itself.

The second improvement remove the additional query by directly injecting
the domain in the query.

With both improvement the _search_activity_state spend 600ms instead of
4000ms for the same search.

The tz optimization can be done as well for _read_group_groupby
the gain are meager from 1200ms to 800ms on the test database

Also improve test with freeze time to make it independent from test
hours as it is crashing around midnight.

Task-4105401

closes odoo/odoo#175927

Related: odoo/upgrade#6358
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>